[Move milestonePanel components to milestoneLibrary dir]
